week3/requirements.txt
pyautogen 
crewai 
langchain 
langgraph 
openai 
python-dotenv 
google-cloud-aiplatform
autogen